Fixed tweakers.net
Comments weren't blocked on tweakers.net anymore. In the current version of the site 'reactieContainer' is an id instead of a class. Instead of blocking reactieContainer, I opted to block all 'reacties' ids, since that will work on a broad range of sites in Dutch and it's a parent of reactieContainer on tweakers.net.